# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for managing visitor registrations with PostgreSQL and JSONB storage.",
        "title": "Visitor Management API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/healthz": {
            "get": {
                "description": "Returns service health status including database connectivity",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/visitors": {
            "get": {
                "description": "Retrieve a paginated list of visitors",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visitors"
                ],
                "summary": "List all visitors",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 20, max: 100)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.PaginatedResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new visitor record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visitors"
                ],
                "summary": "Create a visitor",
                "parameters": [
                    {
                        "description": "Visitor data",
                        "name": "visitor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/visitors/{id}": {
            "get": {
                "description": "Retrieve a visitor's details by their UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visitors"
                ],
                "summary": "Get a visitor by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Visitor UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing visitor's data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visitors"
                ],
                "summary": "Update a visitor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Visitor UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated visitor data",
                        "name": "visitor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a visitor by their UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visitors"
                ],
                "summary": "Delete a visitor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Visitor UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.createRequest": {
            "type": "object",
            "required": [
                "city",
                "country",
                "email",
                "name"
            ],
            "properties": {
                "city": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "CDMX"
                },
                "country": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "Mexico"
                },
                "email": {
                    "type": "string",
                    "example": "miguel@example.com"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "Miguel"
                }
            }
        },
        "response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        }
    }
}